[feat/back] 스크랩완료

+++ ssalog_backend/src/main/java/com/ssalog/service/ScrapServiceImpl.java
@@ -23,13 +23,13 @@ import com.ssalog.repository.ScrapRepository;
 public class ScrapServiceImpl implements ScrapService {
 	@Autowired
 	ScrapRepository scrapRepository;
-	
+
 	@Autowired
 	AccountRepository accountRepository;
-	
+
 	@Autowired
 	PostRepository postRepository;
-	
+
 	@Autowired
 	PostSubRepository postSubRepository;
 	@Override
@@ -37,14 +37,17 @@ public class ScrapServiceImpl implements ScrapService {
 	public String make_scrap(String username, String Scoring ) {
 		Account account = accountRepository.findByUsername(username);
 		Optional<Post> p = postRepository.findById(Scoring);
-		if(p.isPresent() && account != null) {
-			p.get().setLike(p.get().getLike()+1);
-			postRepository.save(p.get());
-			Scrap s = new Scrap();
-			s.setScrap(account);
-			s.setScraped(p.get().getScoring());
-			scrapRepository.save(s);
-			return "success";
+		if(p.isPresent() && account != null ) {
+			Scrap s1 = scrapRepository.findByScrap_usernameAndScraped(username, Scoring);
+			if(s1 == null) {
+				p.get().setLike(p.get().getLike()+1);
+				postRepository.save(p.get());
+				Scrap s = new Scrap();
+				s.setScrap(account);
+				s.setScraped(p.get().getScoring());
+				scrapRepository.save(s);
+				return "success";
+			}
 		}
 		return "fail";
 	}
@@ -53,13 +56,16 @@ public class ScrapServiceImpl implements ScrapService {
 		Account account = accountRepository.findByUsername(username);
 		Optional<Post> p = postRepository.findById(Scoring);
 		if(p.isPresent() && account != null) {
-			p.get().setLike(p.get().getLike()-1);
-			postRepository.save(p.get());
-			Scrap s = new Scrap();
-			s.setScrap(account);
-			s.setScraped(p.get().getScoring());
-			scrapRepository.delete(s);
-			return "success";
+			Scrap s1 = scrapRepository.findByScrap_usernameAndScraped(username, Scoring);
+			if(s1 != null) {
+				p.get().setLike(p.get().getLike()-1);
+				postRepository.save(p.get());
+				Scrap s = new Scrap();
+				s.setScrap(account);
+				s.setScraped(p.get().getScoring());
+				scrapRepository.delete(s);
+				return "success";
+			}
 		}
 		return "fail";
 	}
@@ -70,9 +76,36 @@ public class ScrapServiceImpl implements ScrapService {
 		List<Scrapedpost> list = new ArrayList<>();
 		for(int i=0; i<li.size(); i++) {
 			Scrapedpost p = postSubRepository.findByScoring((li.get(i).getScraped()));
-			//m.put(i, (Scrapedpost)p);
 			list.add(p);
 		}
 		return list;
+	}
+
+	@Override
+	public Boolean is_scraped(String username, String Scoring) {
+		Scrap s = scrapRepository.findByScrap_usernameAndScraped(username, Scoring);
+		if(s != null) {
+			return true;
+		}else {
+			return false;
+		}
+	}
+
+	@Override
+	public long findScrapsu(String nickname) {
+		Account ac = accountRepository.findByNickname(nickname);
+		long cnt = 0;
+		if(ac != null) {
+			List<Post> list = postRepository.findlikesu(ac.getUsername());
+			for(int i=0; i<list.size(); i++) {
+				cnt += list.get(i).getLike();
+			}
+			return cnt;
+		}
+		else {
+			return 0;
+		}
+
+		//System.out.println(postRepository.findlikesu(username).get(0)+ " 입니다!");
 	}
 }
